Fetch the right datastore from vsan-direct datastore URL
In the case of underlying WCP cluster with vsan-direct enabled environment,
volume creation was failing as the correct datastore wasn't being picked up.
The following changes adds a new attribute in the create volume spec to
store the datastore URL selected from the storage pool given in the Create
Volume request. Further, we create a datacenter object from the config and
fetch the datastore object specified by the vsan-direct selected datastore URL
and pass it down to CNS for volume provisioning